DeprecationWarning: `node --inspect --debug-brk` is deprecated
I was receiving `(node:703) [DEP0062] DeprecationWarning: `node --inspect --debug-brk` is deprecated. Please use `node --inspect-brk` instead.` because --debug-brk was being added on like 24, but line 23 already ensures that either --inspect or --inspect-brk will already be passed because code only makes it here IF not null at like 11 and that is only set when matched from line 5 which I've updated in include matching for the legacy --debug-brk and --debug (for those not up to speed yet) and cleaned up the logic/code.